refactor content utils
closes #219
