+++ projects/ngx-d3timeline/src/lib/activity-rectangle/activity-rectangle.ts
@@ -13,3 +13,5 @@ export interface ActivityRectangle {
   strokeWidth: number;
   disableDrag: boolean;
 }
+
+export const activityTitlePadding = 2;
