+++ projects/ngx-d3timeline/src/lib/activity-rectangle/selectors/activity-rectangle-position.selectors.ts
@@ -5,23 +5,24 @@ import {
   selectActivityLateralMargin
 } from '../../options/selectors/options.selectors';
 import { selectResourcePadding } from '../../options/selectors/resource-options.selectors';
-import {
-  getPositionInTimeAxis,
-  getPositionInResourceAxis,
-  getActivityX,
-  getActivityY,
-  getActivityTopLeft,
-  getActivityTransform,
-  getOffsetActivityTopLeft
-} from '../../content/content-utils';
 import { MemoizedSelector } from '../../store-lib/selector/memoized-selector';
-import { Point } from '../../core/point';
+import { Point, translatePoint, pointToTransform } from '../../core/point';
 import { partial3, partial2, partial1 } from '../../core/partial';
+import { BandScale, TimeScale } from '../../scales/scale-types';
+import { PositionedActivity } from '../../activity/positioned-activity';
+import { Orientation } from '../../core/orientation';
 
 const selectGetPositionInTimeAxis = createSelector(
   selectTimeScale,
   partial1(getPositionInTimeAxis)
 );
+
+function getPositionInTimeAxis(
+  timeScale: TimeScale,
+  positionedActivity: PositionedActivity
+): number {
+  return timeScale(positionedActivity.updatedStart);
+}
 
 const selectGetPositionInResourceAxis = createSelector(
   selectBandScale,
@@ -30,12 +31,36 @@ const selectGetPositionInResourceAxis = createSelector(
   partial3(getPositionInResourceAxis)
 );
 
+function getPositionInResourceAxis(
+  bandScale: BandScale,
+  resourcePadding: number,
+  getActivityLateralMargin: (type: string) => number,
+  positionedActivity: PositionedActivity
+): number {
+  return (
+    bandScale(positionedActivity.updatedResource) +
+    resourcePadding +
+    getActivityLateralMargin(positionedActivity.type)
+  );
+}
+
 const selectGetActivityX = createSelector(
   selectTimeOrientation,
   selectGetPositionInResourceAxis,
   selectGetPositionInTimeAxis,
   partial3(getActivityX)
 );
+
+function getActivityX(
+  timeOrientation: Orientation,
+  positionInResourceAxis: (p: PositionedActivity) => number,
+  positionInTimeAxis: (p: PositionedActivity) => number,
+  positionedActivity: PositionedActivity
+): number {
+  return timeOrientation === Orientation.Vertical
+    ? positionInResourceAxis(positionedActivity)
+    : positionInTimeAxis(positionedActivity);
+}
 
 const selectGetActivityY = createSelector(
   selectTimeOrientation,
@@ -44,11 +69,30 @@ const selectGetActivityY = createSelector(
   partial3(getActivityY)
 );
 
+function getActivityY(
+  timeOrientation: Orientation,
+  positionInResourceAxis: (p: PositionedActivity) => number,
+  positionInTimeAxis: (p: PositionedActivity) => number,
+  positionedActivity: PositionedActivity
+): number {
+  return timeOrientation === Orientation.Vertical
+    ? positionInTimeAxis(positionedActivity)
+    : positionInResourceAxis(positionedActivity);
+}
+
 const selectGetActivityTopLeft = createSelector(
   selectGetActivityX,
   selectGetActivityY,
   partial2(getActivityTopLeft)
 );
+
+function getActivityTopLeft(
+  x: (p: PositionedActivity) => number,
+  y: (p: PositionedActivity) => number,
+  positionedActivity: PositionedActivity
+): Point {
+  return { x: x(positionedActivity), y: y(positionedActivity) };
+}
 
 const selectGetOffsetActivityTopLeft = (
   selectOffset: MemoizedSelector<Point>
@@ -58,6 +102,14 @@ const selectGetOffsetActivityTopLeft = (
     selectOffset,
     partial2(getOffsetActivityTopLeft)
   );
+
+export function getOffsetActivityTopLeft(
+  activityTopLeft: (p: PositionedActivity) => Point,
+  offset: Point,
+  positionedActivity: PositionedActivity
+): Point {
+  return translatePoint(activityTopLeft(positionedActivity), offset);
+}
 
 const createSelectTopLeft = (selectOffset?: MemoizedSelector<Point>) =>
   selectOffset
@@ -71,3 +123,10 @@ export const selectGetActivityTransform = (
     createSelectTopLeft(selectOffset),
     partial1(getActivityTransform)
   );
+
+function getActivityTransform(
+  activityTopLeft: (p: PositionedActivity) => Point,
+  positionedActivity: PositionedActivity
+): string {
+  return pointToTransform(activityTopLeft(positionedActivity));
+}
